More changes to get rid of compiler warnings.  This time I'm fixing warnings from GNU -Wall.
Former-commit-id: 1200bdceeee8ecca498a0707c46d1fea2e3a1823

+++ src/libcode/vx_time_series/compute_swinging_door.cc
@@ -36,24 +36,24 @@ bool compute_swinging_door_slopes(const TimeArray &valid_times,
   {
     return false;
   }
-  
+
   // Put the observations into an SDObservation array so we can
   // calculate the ramps.
 
   vector< SDObservation > obs;
-  for (size_t i = 0; i < valid_times.n_elements(); ++i)
+  for (int i = 0; i < valid_times.n_elements(); ++i)
     obs.push_back(SDObservation((time_t)valid_times[i], data_values[i]));
 
   // Calculate the ramps
 
   vector< pair< SDObservation, SDObservation > > ramps;
   if (!compute_swinging_door_ramps(obs, error, ramps)) return false;
-  
+
   // Convert the ramps to slope values
 
   slopes.empty();
-  
-  for (size_t i = 0; i < valid_times.n_elements(); ++i)
+
+  for (int i = 0; i < valid_times.n_elements(); ++i)
   {
     // Check for missing data
 
@@ -73,7 +73,7 @@ bool compute_swinging_door_slopes(const TimeArray &valid_times,
       // Otherwise, the end point of the ramp is assigned the previous slope.
 
       if ( (ramp           == ramps.begin()                &&
-            valid_times[i] == ramp->first.getValidTime())  || 
+            valid_times[i] == ramp->first.getValidTime())  ||
            (valid_times[i] >  ramp->first.getValidTime()   &&
             valid_times[i] <= ramp->second.getValidTime()) )
       {
@@ -104,8 +104,8 @@ bool compute_swinging_door_ramps(const vector< SDObservation > &observations,
 
   // Check for empty list.
 
-  if (observations.size() == 0) return true; 
-  
+  if (observations.size() == 0) return true;
+
   // Save the information for the first point
 
   vector< SDObservation >::const_iterator curr_obs = observations.begin();
@@ -114,34 +114,34 @@ bool compute_swinging_door_ramps(const vector< SDObservation > &observations,
   double start_value = curr_obs->getValue();
   double start_top_value = start_value + error;
   double start_bottom_value = start_value - error;
-  
+
   double top_slope = -99999;
   double bottom_slope = 99999;
-  
+
   SDObservation ramp_start = *curr_obs;
-  
+
   // Loop through the rest of the observations
 
   time_t end_time;
   double end_value;
-  
+
   for (++curr_obs; curr_obs != observations.end(); ++curr_obs)
   {
     // Calculate the needed values for this point
 
     end_time = curr_obs->getValidTime();
     end_value = curr_obs->getValue();
-    
+
     double time_diff = (double)(end_time - start_time);
-    
+
     double curr_top_slope = (end_value - start_top_value) / time_diff;
     double curr_bottom_slope = (end_value - start_bottom_value) / time_diff;
-    
+
     if (top_slope < curr_top_slope)
       top_slope = curr_top_slope;
     if (bottom_slope > curr_bottom_slope)
       bottom_slope = curr_bottom_slope;
-    
+
     // See if we are at the end of the current corridor
 
     if (top_slope > bottom_slope)
@@ -150,7 +150,7 @@ bool compute_swinging_door_ramps(const vector< SDObservation > &observations,
 
       double slope = (top_slope + bottom_slope) / 2.0;
       double end_value = start_value + slope * time_diff;
-      
+
       // Save the end point of the previous corridor
 
       SDObservation ramp_end(end_time, end_value);
@@ -165,19 +165,19 @@ bool compute_swinging_door_ramps(const vector< SDObservation > &observations,
 
       ramp_start = *curr_obs;
       ramp_start.setValue(start_value);
-      
+
       top_slope = -99999;
       bottom_slope = 99999;
     }
   }
-  
+
   // Save the last observation to finish off the last corridor.
   // NOTE: I should probably change this to get the middle of the final
   // doors instead.
 
   SDObservation ramp_end(end_time, end_value);
   ramps.push_back(pair< SDObservation, SDObservation >(ramp_start, ramp_end));
-    
+
   return true;
 }
